// packages/v2/unocss/rules/rule-display.d.ts
export declare const RULES_DISPLAY_UNOCSS: {
    block: () => {
        display: string;
    };
    'inline-block': () => {
        display: string;
    };
    inline: () => {
        display: string;
    };
    'inline-flex': () => {
        display: string;
    };
    table: () => {
        display: string;
    };
    'inline-table': () => {
        display: string;
    };
    'table-caption': () => {
        display: string;
    };
    'table-cell': () => {
        display: string;
    };
    'table-column': () => {
        display: string;
    };
    'table-column-group': () => {
        display: string;
    };
    'table-footer-group': () => {
        display: string;
    };
    'table-header-group': () => {
        display: string;
    };
    'table-row-group': () => {
        display: string;
    };
    'table-row': () => {
        display: string;
    };
    'flow-root': () => {
        display: string;
    };
    contents: () => {
        display: string;
    };
    'list-item': () => {
        display: string;
    };
    none: () => {
        display: string;
    };
    hidden: () => {
        visibility: string;
    };
    invisible: () => {
        visibility: string;
    };
    blind: () => {
        position: string;
        width: string;
        height: string;
        padding: string;
        border: string;
        margin: string;
        'white-space': string;
        overflow: string;
        'clip-path': string;
    };
    'sr-only': () => {
        position: string;
        width: string;
        height: string;
        padding: string;
        border: string;
        margin: string;
        'white-space': string;
        overflow: string;
        'clip-path': string;
    };
    gone: () => {
        position: string;
        width: string;
        height: string;
        padding: string;
        border: string;
        margin: string;
        'white-space': string;
        overflow: string;
        'clip-path': string;
    };
    clip: () => Generator<{
        overflow: string;
        'flex-shrink'?: undefined;
    } | {
        [x: number]: (s: string) => string;
        'flex-shrink': string;
        overflow?: undefined;
    }, void, unknown>;
};
//# sourceMappingURL=rule-display.d.ts.map